Refactor models to more specific directories
