+++ rust/processor/src/db/common/models/event_models/mod.rs
@@ -1,0 +1,1 @@
+pub mod raw_events;
