+++ rust/processor/src/db/common/models/mod.rs
@@ -1,1 +1,2 @@
 pub mod default_models;
+pub mod event_models;
