+++ rust/processor/src/db/parquet/models/event_models/mod.rs
@@ -1,0 +1,1 @@
+pub mod parquet_events;
