+++ rust/processor/src/db/parquet/models/event_models/parquet_events.rs
@@ -2,8 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #![allow(clippy::extra_unused_lifetimes)]
-
-use super::raw_events::{EventConvertible, RawEvent, EVENT_TYPE_MAX_LENGTH};
+use crate::db::common::models::event_models::raw_events::{
+    EventConvertible, RawEvent, EVENT_TYPE_MAX_LENGTH,
+};
 use crate::{
     bq_analytics::generic_parquet_processor::{GetTimeStamp, HasVersion, NamedTable},
     utils::util::truncate_str,
