+++ rust/processor/src/db/parquet/models/mod.rs
@@ -1,1 +1,2 @@
 pub mod default_models;
+pub mod event_models;
