+++ rust/processor/src/db/postgres/models/events_models/events.rs
@@ -8,7 +8,7 @@ use aptos_protos::transaction::v1::Event as EventPB;
 use field_count::FieldCount;
 use serde::{Deserialize, Serialize};
 
-use super::raw_events::{EventConvertible, RawEvent};
+use crate::db::common::models::event_models::raw_events::{EventConvertible, RawEvent};
 
 #[derive(Clone, Debug, Deserialize, FieldCount, Identifiable, Insertable, Serialize)]
 #[diesel(primary_key(transaction_version, event_index))]
