+++ rust/processor/src/db/postgres/models/events_models/mod.rs
@@ -2,5 +2,3 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod events;
-pub mod parquet_events;
-pub mod raw_events;
