+++ rust/sdk-processor/src/parquet_processors/mod.rs
@@ -11,7 +11,7 @@ use async_trait::async_trait;
 use enum_dispatch::enum_dispatch;
 use google_cloud_storage::client::{Client as GCSClient, ClientConfig as GcsClientConfig};
 use parquet::schema::types::Type;
-use processor::db::postgres::models::events_models::parquet_events::Event;
+use processor::db::parquet::models::event_models::parquet_events::Event;
 use processor::{
     db::parquet::models::default_models::{
         parquet_move_modules::MoveModule, parquet_move_resources::MoveResource,
