+++ rust/sdk-processor/src/parquet_processors/parquet_events_processor.rs
@@ -30,7 +30,7 @@ use aptos_indexer_processor_sdk::{
 use parquet::schema::types::Type;
 use processor::{
     bq_analytics::generic_parquet_processor::HasParquetSchema,
-    db::postgres::models::events_models::parquet_events::Event as EventPQ,
+    db::parquet::models::event_models::parquet_events::Event as EventPQ,
 };
 use std::{collections::HashMap, sync::Arc};
 use tracing::{debug, info};
